+++ src/app/modules/demo/decorator/decorator.component.ts
@@ -4,10 +4,13 @@ import {ngxLocalStorage} from "../../ngx-localstorage/lib/decorators";
 @Component({
   selector: 'app-decorator',
   templateUrl: './decorator.component.html',
-  styleUrls: ['./decorator.component.less']
+  styleUrls: ['../base/base.scss']
 })
 export class DecoratorComponent {
 
-  @ngxLocalStorage({key: 'cbox1', prefix: 'demo'})
-  boundCheckBox: boolean;
+  @ngxLocalStorage({key: 'cbox1', prefix: 'demo', nullTransformer: () => false})
+  boundBoolean: boolean;
+
+  @ngxLocalStorage({key: 'txt1', prefix: 'demo'})
+  boundText: string;
 }