+++ src/app/modules/demo/demo.module.ts
@@ -1,12 +1,18 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {DecoratorComponent} from './decorator/decorator.component';
+
+import {MatCardModule} from '@angular/material';
+
 import {NgxLocalStorageModule} from "../ngx-localstorage/ngx-localstorage.module";
+
+import {DecoratorComponent} from './decorator/decorator.component';
 import {DirectiveComponent} from './directive/directive.component';
 
 @NgModule({
   imports: [
     CommonModule,
+
+    MatCardModule,
 
     NgxLocalStorageModule.forRoot()
   ],