+++ src/app/modules/ngx-localstorage/lib/decorators.ts
@@ -1,6 +1,6 @@
 /**
-* Created by HOFFM59 on 22.05.2017.
-*/
+ * Created by bohoffi on 22.05.2017.
+ */
 import {DecoratorOpts} from './interfaces';
 import {LocalStorageService} from './services/local-storage.service';
 import {StorageEventService} from './services/storage-event.service';
@@ -16,24 +16,22 @@ export function ngxLocalStorage(options?: DecoratorOpts) {
 
     const eventService: StorageEventService = new StorageEventService();
     eventService.stream
-      // TODO: filter should be more accurate
+    // TODO: filter should be more accurate
       .filter((ev: StorageEvent) => ev.key.indexOf(key) >= 0)
       .subscribe((ev: StorageEvent) => {
-        //target[propertyDescription] = ev.newValue;
         if (!!ev.newValue && typeof ev.newValue === 'string') {
           if (ev.newValue !== 'null') {
-            //setProperty(this._valuePath.length ? this._valuePath : ['value'], ev.newValue, this._element.nativeElement);
             target[propertyDescription] = ev.newValue;
           } else {
-            //setProperty(this._valuePath.length ? this._valuePath : ['value'], '', this._element.nativeElement);
-            target[propertyDescription] = '';
+            target[propertyDescription] = !!options.nullTransformer ? options.nullTransformer() : null;
           }
         }
       });
 
     Object.defineProperty(target, propertyDescription, {
       get: function () {
-        return service.get(key);
+        const storageValue = service.get(key);
+        return storageValue == null && !!options.nullTransformer ? options.nullTransformer() : storageValue;
       },
       set: function (value: any) {
         service.set(key, value)