+++ src/app/modules/ngx-localstorage/lib/services/index.ts
@@ -2,3 +2,4 @@
  * Created by bohoffi on 04.04.2017.
  */
 export * from './local-storage.service';
+export * from './storage-event.service';