+++ src/app/modules/ngx-localstorage/ngx-localstorage.module.ts
@@ -53,5 +53,5 @@ export class NgxLocalStorageModule {
   }
 }
 
-export {LocalStorageService} from './lib/index';
+export {LocalStorageService, StorageEventService} from './lib/index';
 export {ngxLocalStorage} from './lib/decorators';